Install breeze , Register , login and protected route

+++ app/Http/Controllers/API/V1/PostController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Resources\PostResource;
 use App\Models\Post;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class PostController extends Controller
 {
@@ -14,7 +15,10 @@ class PostController extends Controller
      */
     public function index()
     {
-        return PostResource::collection(Post::with('author')->paginate(2));
+        $user = request()->user();
+        $posts = $user->posts()->with('author')->paginate();
+
+        return PostResource::collection($posts);
     }
 
     /**
@@ -27,7 +31,7 @@ class PostController extends Controller
             'body' => 'required|string|min:2',
         ]);
 
-        $data['author_id'] = 1;
+        $data['author_id'] = auth()->id();
 
         $post = Post::create($data);
 
@@ -40,6 +44,8 @@ class PostController extends Controller
      */
     public function show(Post $post)
     {
+        abort_if(Auth::id() != $post->author_id, 403, 'Access Forbidden');
+
         return new PostResource($post);
     }
 
@@ -48,6 +54,9 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
+        abort_if(Auth::id() != $post->author_id, 403, 'Access Forbidden');
+        
+
         $data = $request->validate([
             'title' => 'required|string|min:2',
             'body' => 'required|string|min:2',
@@ -63,6 +72,7 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+         abort_if(Auth::id() != $post->author_id,403,'Access Forbidden');
         $post->delete();
 
         return response()->noContent();
